[refactor] Move login、signup、logout、resetPassword from auth.ts to userinfo
